Simplify path management with a better _OutputPath abstraction
There is no need to keep a separate image_out_dir since images are
always written to the same path as the tests. Things are much simpler
if we use an _OutputPath class to stores both the element and the
offscreen paths. We can then call sub_path(...) on an _OutputPath
to create a sub-path for both element and offscreen.

Bug: 40257559
Change-Id: I5361f4aaad29fc43ed2b3b3a839a1f5fa36a4a11