Remove child content selector from StylableSelect UA rules
This patch removes the :-internal-select-author-datalist and
:-internal-select-author-button pseudoclasses. Based on the standards
discussions, it seems like we will always render the potentially
problematic select UA rules instead of trying to detect whether the new
content model is used or if the computed value of appearance is the new
value. The only problematic styles are the select's border and
background-color, which are probably not a big deal.

Bug: 1511354
Change-Id: I9261183e799c1df57ca2737acee6966137bafeae