Fenced frame: fix CSP tests and CSP check in can load API.
The navigator.canLoadAdAuctionFencedFrame() API currently doesn't
properly handle the case when calling it in a document whose content
security policy is "fenced-frame-src: https:". This is because the API
isn't explicitly checking for that specific case. This CL modifies the
function to match the checks in `CheckCSPSource()` in `csp_source.cc`.

This would've been caught by the CSP WPTs, but those have a bug that
makes them ineffective in catching corner cases. When adding a CSP with
the <meta> tag, that is a strictly additive operation, and removing the
tag will have no effect on the document's CSP. Since the same page is
re-used through all the promise_tests in the file, each subsequent test
will load with all of the CSPs set in the prior tests. This CL fixes
that by having each test create a subframe that gets the CSP, and
running the test in that subframe.

Change-Id: Iea3238cc2b515976468f957ae808277df8c5adc9